dogstatsd: introduction of the OriginTracker
The DogStatsD origin tracker is used to count packets, bytes and tags
count per
origin (when the information is available).

Only implemented in the UDS listener for now, the goal is to have
visibility on
whose client is the noisy one when the traffic isn't balanced between
apps sending
DogStatsD messages to the Agent.